requirements.txt
requests
flask
transformers
datasets
sentence-transformers
wandb
tqdm
pyyaml
zstandard